Output settings for multi object ID index Control input values on validity Issue SVO-3902

+++ SVObserver/SVOInspectionList.cpp
@@ -128,7 +128,7 @@ bool SVOInspectionList::SetObjectIdIndex(LPCTSTR InspectionName, DWORD objectIdI
 
 	if (nullptr != pObj)
 	{
-		pObj->SetObjectIdIndex(objectIdIndex);
+		pObj->SetObjectIdIndex(static_cast<long> (objectIdIndex));
 		return true;
 	}
